Removed Factory from CloudProvider (#460)
* Removed Factory from CloudProvider

* changed api to cloudProvider
