--- pkg/cloudprovider/fake/factory.go
@@ -1,30 +1,0 @@
-/*
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package fake
-
-import (
-	provisioning "github.com/awslabs/karpenter/pkg/apis/provisioning/v1alpha1"
-	"github.com/awslabs/karpenter/pkg/cloudprovider"
-)
-
-type Factory struct{}
-
-func (f *Factory) CapacityFor(provisioner *provisioning.Provisioner) cloudprovider.Capacity {
-	return &Capacity{}
-}
-
-func (f *Factory) TerminateFor() cloudprovider.Termination {
-	return &Termination{}
-}
