+++ pkg/controllers/provisioning/v1alpha1/allocation/filter.go
@@ -19,7 +19,6 @@ import (
 	"fmt"
 
 	"github.com/awslabs/karpenter/pkg/apis/provisioning/v1alpha1"
-	"github.com/awslabs/karpenter/pkg/cloudprovider"
 	"github.com/awslabs/karpenter/pkg/utils/functional"
 	"github.com/awslabs/karpenter/pkg/utils/pod"
 	"github.com/awslabs/karpenter/pkg/utils/ptr"
@@ -30,8 +29,7 @@ import (
 )
 
 type Filter struct {
-	kubeClient    client.Client
-	cloudProvider cloudprovider.Factory
+	kubeClient client.Client
 }
 
 func (f *Filter) GetProvisionablePods(ctx context.Context, provisioner *v1alpha1.Provisioner) ([]*v1.Pod, error) {
