+++ pkg/controllers/provisioning/v1alpha1/allocation/suite_test.go
@@ -44,7 +44,7 @@ func TestAPIs(t *testing.T) {
 
 var controller *Controller
 var env = test.NewEnvironment(func(e *test.Environment) {
-	cloudProvider := &fake.Factory{}
+	cloudProvider := &fake.CloudProvider{}
 	registry.RegisterOrDie(cloudProvider)
 	controller = NewController(
 		e.Manager.GetClient(),
