+++ pkg/controllers/provisioning/v1alpha1/reallocation/controller.go
@@ -35,7 +35,7 @@ import (
 // Controller for the resource
 type Controller struct {
 	utilization   *Utilization
-	cloudProvider cloudprovider.Factory
+	cloudProvider cloudprovider.CloudProvider
 }
 
 // For returns the resource this controller is for.
@@ -57,7 +57,7 @@ func (c *Controller) Name() string {
 }
 
 // NewController constructs a controller instance
-func NewController(kubeClient client.Client, coreV1Client corev1.CoreV1Interface, cloudProvider cloudprovider.Factory) *Controller {
+func NewController(kubeClient client.Client, coreV1Client corev1.CoreV1Interface, cloudProvider cloudprovider.CloudProvider) *Controller {
 	return &Controller{
 		utilization:   &Utilization{kubeClient: kubeClient},
 		cloudProvider: cloudProvider,
