+++ pkg/controllers/terminating/v1alpha1/suite_test.go
@@ -39,7 +39,7 @@ func TestAPIs(t *testing.T) {
 
 var controller *Controller
 var env = test.NewEnvironment(func(e *test.Environment) {
-	cloudProvider := &fake.Factory{}
+	cloudProvider := &fake.CloudProvider{}
 	registry.RegisterOrDie(cloudProvider)
 	controller = NewController(
 		e.Manager.GetClient(),
